Make index page template-based
This content is based on chapters 2 thru 5 of the Mega Tutorial,
specifically the additions to the index page and the base template.
The login page and form are split into a separate commit to follow.

https://blog.miguelgrinberg.com/post/the-flask-mega-tutorial-part-ii-templates

**Deviations from the source material**

The dummy "posts" on the index page have been replaced with a db query
for the real posts. Right now it's all of them, but I expect the index
page will be rewritten before the number of messages gets excessive.

Also, a link to the `/user` endpoint has been added to the header in
the base template, in order to tie the different parts of this flask
app together.

**Motivation**

These files are being added because they indirectly support the login
code. It was possible to rewrite the login page to not use base
template, and the code to redirect elsewhere after logging in, but
that made it harder to follow the later chapters. Plus the base
template is a nice feature to have.