Add login page and form, plus initial logic
This content is directly based on Chapter 3 of the Mega Tutorial.

https://blog.miguelgrinberg.com/post/the-flask-mega-tutorial-part-iii-web-forms

A lot of stuff is added, but since my goal is to provide the code that
Chapter 5 uses as its foundation, it seems reasonable to squash these
changes together.

**Core Concepts**

This commit adds Flask-WTF, which is a wrapper for WTForms. This
package enables us to represent a form as a Python object, and makes it
easier to describe the content and validation logic. Once the form has
been submitted, an instance of this class gets created and filled with
the sent data. In these ways, it's a lot like SQLAlchemy.

Note that the presentation of the form is handled separately from the
definition and logic. The elements of the form object provide their
individual html snippets, but the layout of these on the page is
handled by login page template.

When the submission doesn't pass validation, the `errors` properties
get populated on the problematic fields. The template then has blocks
for displaying these.

Finally, in routes.py we have the view function for the new endpoint.
This code decides what should happen in response to the valid/invalid
submission data.